Organise storage docs for mdbook
